+++ crates/ruma-events/CHANGELOG.md
@@ -75,6 +75,9 @@ Improvements:
   now implements `Ord` using the aforementioned trait.
 - With the `compat-lax-room-create-deser` cargo feature, the `predecessor` field of
   `RoomCreateEventContent` is ignored during deserialization if it has an invalid format.
+- With the `compat-lax-room-topic-deser` cargo feature, the `topic_block` field of
+  `RoomTopicEventContent` (named `m.topic` in the JSON source) is ignored during deserialization if
+  it has an invalid format.
    
 # 0.30.3